Rearranging structure/stress macros
